+++ bspline_tools/bsplineNkernel.m
@@ -1,6 +1,0 @@
-function b = bsplineNkernel(k,n,m)
-% centered B-spline kernel of order n, expansion factor m
-
-b = bsplineN(k/m,n);
-
-end