+++ f_to_srvf.m
@@ -11,7 +11,7 @@ function q = f_to_srvf(f,time,smooth,spl, parallel)
 % f: matrix of functions
 % time: vector of time samples
 % smooth: use smoothing splines (default: true)
-% spl: use b-spline computation (old behavior)
+% spl: use interpolating cubic spline derivative
 % paralell: compute in parallel
 %
 % Output:
@@ -61,20 +61,15 @@ if smooth
     end
     q = fy./sqrt(abs(fy)+eps);
 elseif spl
+    tt = time(:);
     fy = zeros(M,N);
     if parallel
         parfor ii = 1:N
-            y = Bspline(f(:,ii),3);
-            ydiff = diff(y);
-            fy(:,ii) = ydiff(1:length(time))/binsize;
-            f(:,ii) = y(1:length(time));
+            fy(:,ii) = fnval(fnder(csapi(tt, f(:,ii))), tt);
         end
     else
         for ii = 1:N
-            y = Bspline(f(:,ii),3);
-            ydiff = diff(y);
-            fy(:,ii) = ydiff(1:length(time))/binsize;
-            f(:,ii) = y(1:length(time));
+            fy(:,ii) = fnval(fnder(csapi(tt, f(:,ii))), tt);
         end
     end
     q = fy./sqrt(abs(fy)+eps);